[#88] :truck: rename: 일기장 수정/생성 페이지 일기장 색상 atom 공통 사용에 따라 디렉토리명 변경
