+++ client/src/components/new-book/DiaryIcon.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import tw from "tailwind-styled-components";
 import { useRecoilValue } from "recoil";
 
-import { selectedColorAtom } from "../../recoil/new-book";
+import { selectedColorAtom } from "../../recoil/book-color";
 
 const DiaryImg = tw.img`
     mx-auto pt-[9vh] h-[27vh] md:h-[25vh]
